requirements.txt
census==0.8.13
Django==3.0.3
django-cors-headers==3.1.1
django-extensions==2.1.9
django-filter==2.1.0
django-leaflet==0.24.0
django-model-utils==3.2.0
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
djangorestframework==3.11.0
psycopg2-binary==2.8.2
pygraphviz==1.5
requests==2.21.0
uncertainties==3.0.3
django-grappelli
djangorestframework-camel-case
django-nested-admin